py_script/requirements.txt
pyserial
aiohttp
asyncio
